feature: move timber-settings activity to application
